Move FileSystemManager to utils
