--- core/src/main/java/org/polypheny/db/filemanagement/FileSystemManager.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2019-2020 The Polypheny Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.polypheny.db.filemanagement;
-
-
-import java.io.File;
-
-
-public abstract class FileSystemManager {
-
-    public abstract File registerDataFolder( String path );
-
-    public abstract File registerNewFile( String pathToFile );
-
-    public abstract File registerNewFile( File path, String pathToFile );
-}
